pref(excel): 优化代码结构
